examples and test
