+++ Backpack-SwiftUI/MapMarker/Classes/BPKPriceMapMarker.swift
@@ -16,11 +16,10 @@
  * limitations under the License.
  */
 
-import MapKit
 import SwiftUI
 
 public struct BPKPriceMapMarker: View {
-    enum State {
+    public enum State {
         case `default`
         case focused
         case viewed
@@ -67,20 +66,22 @@ public struct BPKPriceMapMarker: View {
         }
     }
     
-    let state: State
-    let price: String
+    private let state: State
+    private let price: String
+    
+    public init(state: State, price: String) {
+        self.state = state
+        self.price = price
+    }
     
     public var body: some View {
-        BPKFlareViewMap(
-            size: .small,
-            direction: .bottom,
-            content: content
-        )
-        .shadow(.sm)
+        content
+            .clipShape(MarkerShape())
+            .shadow(.sm)
     }
     
     @ViewBuilder
-    func content() -> some View {
+    private var content: some View {
         if case .focused = state {
             VStack(spacing: BPKSpacing.none) {
                 BPKText(price, style: state.fontStyle)
@@ -107,223 +108,13 @@ public struct BPKPriceMapMarker: View {
     }
 }
 
-public struct BPKPointerMapMarker: View {
-    
-    public var body: some View {
-        Circle()
-            .stroke(Color(.surfaceDefaultColor), lineWidth: 2)
-            .background(Circle().foregroundColor(.coreAccentColor))
-            .frame(width: .base, height: .base)
-            
-    }
-}
-
-public struct BPKPoiMapMarker: View {
-    enum State {
-        case `default`
-        case focused
-        case disabled
-        
-        var foregroundColor: BPKColor {
-            switch self {
-            case .default:
-                return .textPrimaryInverseColor
-            case .focused:
-                return .coreAccentColor
-            case .disabled:
-                return .textDisabledColor
-            }
-        }
-        
-        var backgroundColor: BPKColor {
-            switch self {
-            case .default:
-                return .coreAccentColor
-            case .focused:
-                return .surfaceDefaultColor
-            case .disabled:
-                return .surfaceDefaultColor
-            }
-        }
-    }
-    
-    let state: State
-    let icon: BPKIcon
-    
-    public var body: some View {
-        ZStack {
-            Color(state.backgroundColor)
-                .frame(width: 26, height: 32)
-                .clipShape(BPKPOIMapMarkerShape())
-                
-            BPKIconView(icon)
-                .foregroundColor(state.foregroundColor)
-                .padding(.bottom, 6)
-        }
-    }
-}
-
-/// A view that clips its content to a shape with a flare at one end.
-struct BPKPOIMapMarkerShape: Shape {
-    /// Creates a path with a flare at one end.
-    func path(in rect: CGRect) -> Path {
-        let size = rect.size
-        return Path { path in
-            path.addRelativeArc(
-                center: .init(x: size.width / 2, y: size.width / 2),
-                radius: size.width / 2,
-                startAngle: .degrees(30),
-                delta: .degrees(-240)
-            )
-
-            let courvePoint: CGFloat = 3
-            let startPoint = CGPoint(x: size.width / 2, y: size.height - courvePoint)
-
-            path.addRelativeArc(
-                center: startPoint,
-                radius: courvePoint,
-                startAngle: .degrees(120),
-                delta: .degrees(-60)
-            )
-        }
-    }
-}
-
-/// A view that clips its content to a shape with a flare at one end.
-struct BPKFlareViewMap<Content: View>: View {
-    private let size: BPKFlareSize
-    private let direction: BPKFlareDirection
-    @ViewBuilder private let content: Content
-    
-    /// Creates a view that clips its content to a shape with a flare at one end.
-    /// - Parameters:
-    ///   - size: The size of the flare. Default is `.medium`.
-    ///   - roundedCorners: Whether to round the corners of the shape. Default is `true`.
-    ///   - direction: The direction of the flare. Default is `.bottom`.
-    ///   - content: The content to clip to the shape.
-    public init(
-        size: BPKFlareSize = .medium,
-        direction: BPKFlareDirection = .bottom,
-        content: () -> Content
-    ) {
-        self.size = size
-        self.direction = direction
-        self.content = content()
-    }
-    
-    var body: some View {
-        content
-            .clipShape(FlarePath())
-    }
-    
-    struct FlarePath: Shape {
-        let flareSize = CGSize(width: 16, height: 6)
-        let cornerRadius = BPKSpacing.sm
-        
-        func path(in rect: CGRect) -> Path {
-            let size = rect.size
-            return Path { path in
-                path.addRoundedRect(
-                    in: CGRect(
-                        origin: .zero,
-                        size: CGSize(width: size.width, height: size.height - flareSize.height)
-                    ),
-                    cornerSize: CGSize(width: cornerRadius.value, height: cornerRadius.value))
-                appendFlare(
-                    to: &path,
-                    startPoint: CGPoint(x: size.width / 2, y: size.height),
-                    flareSize: flareSize
-                )
-            }
-        }
-        
-        private func appendFlare(to path: inout Path, startPoint: CGPoint, flareSize: CGSize) {
-            let startPointY = startPoint.y - flareSize.height
-            let centerPoint = startPoint.x
-            path.move(to: CGPoint(x: centerPoint - flareSize.width / 2, y: startPointY))
-            
-            let courvePoint: CGFloat = 2
-            path.addLine(to: CGPoint(x: centerPoint - courvePoint, y: startPoint.y - courvePoint / 2))
-            
-            let center = CGPoint(x: centerPoint + courvePoint, y: startPoint.y - courvePoint / 2)
-            let curveCenter1 = CGPoint(x: centerPoint - courvePoint / 2, y: startPointY + flareSize.height)
-            let curveCenter2 = CGPoint(x: centerPoint + courvePoint / 2, y: startPointY + flareSize.height)
-            path.addCurve(to: center, control1: curveCenter1, control2: curveCenter2)
-            
-            path.addLine(to: CGPoint(x: centerPoint + flareSize.width / 2, y: startPointY))
-        }
-    }
-}
-
 struct BPKPriceMapMarker_Previews: PreviewProvider {
-    struct Annotation: Identifiable {
-        // swiftlint:disable nesting
-        enum Marker {
-            case price(String, BPKPriceMapMarker.State)
-            case poi(BPKIcon, BPKPoiMapMarker.State)
-            case pointer
-        }
-        let id = UUID()
-        let marker: Marker
-        let coordinate: CLLocationCoordinate2D
-    }
-    
     static var previews: some View {
         VStack {
-            BPKPoiMapMarker(state: .default, icon: .landmark)
-            BPKPoiMapMarker(state: .default, icon: .landmark)
-            BPKPoiMapMarker(state: .default, icon: .landmark)
-            Spacer()
-        }
-        Map(
-            coordinateRegion: .constant(.init(
-                center: .init(latitude: 51.5, longitude: -0.1),
-                span: .init(latitudeDelta: 0.2, longitudeDelta: 0.2)
-            )),
-            annotationItems: [
-                Annotation(
-                    marker: .price("£200", .default),
-                    coordinate: .init(latitude: 51.55, longitude: -0.1)
-                ),
-                Annotation(
-                    marker: .price("£200", .focused),
-                    coordinate: .init(latitude: 51.53, longitude: -0.1)
-                ),
-                Annotation(
-                    marker: .price("£200", .viewed),
-                    coordinate: .init(latitude: 51.51, longitude: -0.1)
-                ),
-                Annotation(
-                    marker: .price("Sold out", .disabled),
-                    coordinate: .init(latitude: 51.49, longitude: -0.1)
-                ),
-                Annotation(
-                    marker: .pointer,
-                    coordinate: .init(latitude: 51.47, longitude: -0.1)
-                ),
-                Annotation(
-                    marker: .poi(.landmark, .default),
-                    coordinate: .init(latitude: 51.55, longitude: -0.04)
-                ),
-                Annotation(
-                    marker: .poi(.landmark, .focused),
-                    coordinate: .init(latitude: 51.53, longitude: -0.04)
-                ),
-                Annotation(
-                    marker: .poi(.landmark, .disabled),
-                    coordinate: .init(latitude: 51.51, longitude: -0.04)
-                )
-            ]) { item in
-                MapAnnotation(coordinate: item.coordinate) {
-                    switch item.marker {
-                    case .price(let price, let state):
-                        BPKPriceMapMarker(state: state, price: price)
-                    case .poi(let icon, let state):
-                        BPKPoiMapMarker(state: state, icon: icon)
-                    case .pointer:
-                        BPKPointerMapMarker()
-                    }
-                }
+            BPKPriceMapMarker(state: .default, price: "£200")
+            BPKPriceMapMarker(state: .focused, price: "£200")
+            BPKPriceMapMarker(state: .viewed, price: "£200")
+            BPKPriceMapMarker(state: .disabled, price: "Sould out")
         }
     }
 }
